Move Container-related classes to specific namespace
